requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
pyYAML
xgboost
catboost
lightgbm
joblib
evidently
boto3
mlflow
dvc
pyYAML
Flask


# dev
flake8
black


